+++ MyShuffle/vectors.h
@@ -47,7 +47,7 @@ public:
 	const mpc_int& at(size_t pos) const;
 	mpc_int& back(size_t line);
 
-	void resize(int newnum, int newlen);
+	void resize(size_t newnum, size_t newlen);
 	void push_back(const mpc_int* val);
 
 	void clear();
@@ -308,7 +308,7 @@ inline mpc_int& vectors<mpc_int>::back(size_t line)
 }
 
 template<typename mpc_int>
-inline void vectors<mpc_int>::resize(int newnum, int newlen)
+inline void vectors<mpc_int>::resize(size_t newnum, size_t newlen)
 {
 	num = newnum;
 	len = newlen;
@@ -324,7 +324,7 @@ inline void vectors<mpc_int>::push_back(const mpc_int* val) {
 		std::cerr << "vectors::push_back: len == 0." << std::endl;
 		throw "vectors::push_back: len == 0.";
 	}
-	for (int i(0); i != len; ++i) {
+	for (size_t i(0); i != len; ++i) {
 		vec.push_back(val[i]);
 	}
 	++num;
@@ -390,9 +390,9 @@ inline size_t vectors<mpc_int>::size() const
 template <typename mpc_int>
 std::ostream& operator<<(std::ostream& os, const vectors<mpc_int>& v) {
 	os << "Vectors " << &v << " {" << std::endl;
-	for (int i(0); i != v.num; ++i) {
+	for (size_t i(0); i != v.num; ++i) {
 		os << "\t";
-		for (int j(0); j != v.len; ++j) {
+		for (size_t j(0); j != v.len; ++j) {
 			os << v[i][j] << " ";
 
         }
